fix(store): adjust behavior of action and state stream within the store
The action stream does not need to start with an initial action so it is changed to an Observable
derived from a Subject, and since it uses a Subject internally the .share call is unecessary. The
state stream doesn't just have to replay the last value, it needs to always have the newest value
returned. This is accomplished by multicasting with a BehaviorSubject, using refCout for
subscription management.